organize i32 and i64 codes
